requirements.txt
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
arch==5.3.1
asgiref==3.6.0
astroid==2.3.3
async-timeout==4.0.2
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
backtrader==1.9.76.123
bidict==0.22.1
billiard==3.6.4.0
bleach==3.1.1
Brotli==1.0.9
cachetools==5.3.0
celery==5.2.7
certifi==2019.11.28

chardet==3.0.4
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
contourpy==1.1.1

cycler==0.10.0
Cython==0.29.33
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.6.0
dependency-injector==4.41.0
distlib==0.3.7
dnspython==2.2.1
elasticsearch==7.5.1
entrypoints==0.3
eventlet==0.24.1
fastapi==0.94.1
fastapi-socketio==0.0.9
filelock==3.12.2
Flask==1.1.2
Flask-Compress==1.13
Flask-SocketIO==4.3.0
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2024.5.0
gevent==20.5.2
gitdb==4.0.10
GitPython==3.1.24
greenlet==0.4.15
grpcio==1.57.0
gunicorn==20.0.4
h11==0.14.0
hiredis==2.0.0
humanfriendly==10.0
idna==2.9
importlib-metadata==6.0.0
importlib_resources==6.4.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kafka-python==2.0.2
kiwisolver==1.1.0
kombu==5.2.4
lazy-object-proxy==1.4.3
line-profiler==4.0.3
llvmlite==0.39.1
MarkupSafe==1.1.1
matplotlib==3.7.5
mccabe==0.6.1
mistune==0.8.4
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
nbconvert==5.6.1
nbformat==5.0.4
networkx==3.1
notebook==6.0.3
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
orjson==3.8.8
packaging==23.0
pandas==1.3.4
pandocfilters==1.4.2
parso==0.6.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
pillow==10.3.0
platformdirs==3.10.0
plotly==5.10.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
property-cached==1.6.4
protobuf==4.24.0
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.6.0
py-cpuinfo==9.0.0
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.10.4
Pygments==2.6.1
pylint==2.4.4
pymongo==3.10.1

pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-multipart==0.0.6
python-socketio==5.5.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==19.0.0
ray==2.3.0
redis==4.5.1
requests==2.23.0
retrying==1.3.4
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.13.2
Send2Trash==1.5.0
sentry-sdk==0.10.2
six==1.14.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.1.0
SQLAlchemy==1.4.46
starlette==0.26.1
statsmodels==0.13.5
sympy==1.12

tabulate==0.9.0
tenacity==8.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tornado==6.0.4
tqdm==4.66.4
traitlets==4.3.3
triton==2.3.0
typing_extensions==4.11.0
ultralytics==8.2.19
urllib3==1.25.7
uvicorn==0.21.0
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.1.8
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
wrapt==1.11.2
zipp==3.11.0
zope.event==5.0
zope.interface==6.0
